Move `density_iw` to submodule to avoid cyclic import
